LayoutTests/platform/chromium/media/media-volume-slider-hit-test-expected.txt

Clicking middle of volume slider should set the value to middle.

PASS volumeSlider.value is "5"
PASS successfullyParsed is true

TEST COMPLETE
